clients/multi: move QtTab to sailui
