+++ packages/drivechain_client/lib/pages/root_page.dart
@@ -3,7 +3,7 @@ import 'package:dotted_border/dotted_border.dart';
 import 'package:drivechain_client/routing/router.dart';
 import 'package:drivechain_client/util/color.dart';
 import 'package:flutter/material.dart';
-import 'package:sail_ui/sail_ui.dart';
+import 'package:sail_ui/widgets/nav/top_nav.dart';
 
 @RoutePage()
 class RootPage extends StatelessWidget {
@@ -75,8 +75,7 @@ class RootPage extends StatelessWidget {
                             ..moveTo(size.width, size.height)
                             ..lineTo(size.width, 0),
                           child: Padding(
-                            padding:
-                                const EdgeInsets.symmetric(horizontal: 4.0),
+                            padding: const EdgeInsets.symmetric(horizontal: 4.0),
                             child: QtTab(
                               icon: const Icon(Icons.link),
                               label: 'Sidechains',
@@ -104,48 +103,6 @@ class RootPage extends StatelessWidget {
           ),
         );
       },
-    );
-  }
-}
-
-// TODO: Move this to sail_ui if wanted
-class QtTab extends StatelessWidget {
-  final String label;
-  final Widget icon;
-  final bool active;
-  final VoidCallback onTap;
-
-  const QtTab({
-    super.key,
-    required this.label,
-    required this.icon,
-    required this.active,
-    required this.onTap,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return MouseRegion(
-      cursor: SystemMouseCursors.click,
-      child: GestureDetector(
-        onTap: onTap,
-        child: AnimatedContainer(
-          duration: const Duration(milliseconds: 50),
-          padding: const EdgeInsets.all(8),
-          decoration: BoxDecoration(
-            color: active ? Colors.grey.withOpacity(0.2) : Colors.transparent,
-            borderRadius: BorderRadius.circular(4),
-          ),
-          child: Row(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              icon,
-              const SizedBox(width: 2),
-              SailText.primary12(label),
-            ],
-          ),
-        ),
-      ),
     );
   }
 }
